backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pypdf==5.9.0
Pillow==11.2.1
python-docx==1.1.2
pymupdf==1.25.1
redis==6.2.0
pdf2docx==0.5.8
